Adding empires2x1p1 module.

+++ aoe2m-io/main/java/net/aegistudio/aoe2m/io/FieldInputStream.java
@@ -42,7 +42,7 @@ public class FieldInputStream extends InputStream {
 	}
 	
 	public int readUnsigned16() throws IOException {
-		return (short) readBytes(2, false);
+		return (int) readBytes(2, false);
 	}
 	
 	public short readSigned16() throws IOException {
